test(testkit): move client coverage and reject invalid states
