+++ tests/integration/tests/testkit.rs
@@ -3,6 +3,9 @@
 #[path = "testkit/auth_integration.rs"]
 mod auth_integration;
 
+#[path = "testkit/client_integration.rs"]
+mod client_integration;
+
 #[path = "testkit/static_files_integration.rs"]
 mod static_files_integration;
